requirements.txt
alembic==0.9.6
aniso8601==8.1.0
bcrypt==3.1.4
cffi==1.14.4
click==6.7
coverage==4.4.2
dnspython==2.0.0
email-validator==1.1.2
enum-compat==0.0.3
enum34==1.1.10
eventlet==0.21.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
flask-mongoengine==1.0.0
flask-restplus==0.13.0
flask-restx==0.2.0
Flask-Script==2.0.6
Flask-Testing==0.7.1
Flask-WTF==0.14.3
greenlet==0.4.17
gunicorn==19.7.1
idna==2.10
itsdangerous==0.24
Jinja2==2.11.2
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.1.1
mongoengine==0.21.0
psycopg2-binary==2.8.5
pycparser==2.18
PyJWT==1.5.3
pymongo==3.11.2
python-dateutil==2.6.1
python-editor==1.0.4
pytz==2017.3
selenium==3.8.1
six==1.11.0
SQLAlchemy==1.3.20
Werkzeug==0.15.5
WTForms==2.3.3
